+++ packages/lex-cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atproto/lex-cli
 
+## 0.10.11
+
+### Patch Changes
+
+- Updated dependencies [[`6af9f14`](https://github.com/bluesky-social/atproto/commit/6af9f141d18c1f41a04828af8256567d5f17a35e)]:
+  - @atproto/syntax@0.7.5
+  - @atproto/lexicon@0.7.12
+
 ## 0.10.10
 
 ### Patch Changes