+++ packages/lex/lex-server/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @atproto/lex-server
 
+## 0.1.16
+
+### Patch Changes
+
+- Updated dependencies [[`ec29fcd`](https://github.com/bluesky-social/atproto/commit/ec29fcdb4e0ce098fb1473feebe876a94b1c8e33)]:
+  - @atproto/lex-client@0.3.5
+
+## 0.1.15
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @atproto/lex-client@0.3.4
+  - @atproto/lex-schema@0.2.6
+
 ## 0.1.14
 
 ### Patch Changes